Move redux hooks
