+++ apps/dashboard/src/features/notifications/NotificationsQueue.tsx
@@ -1,7 +1,7 @@
 import { memo, useCallback } from 'react'
 
 import NotificationCard from '@/src/entities/notifications/ui/NotificationCard'
-import { useAppSelector, useAppDispatch } from '@/src/shared/model/hooks'
+import { useAppSelector, useAppDispatch } from '@/src/shared/hooks'
 import { removeNotification } from '@/src/shared/model/notificationsSlice'
 import { Notification } from '@/src/shared/model/Notification'
