+++ apps/dashboard/src/shared/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './usePushNotification'
+export * from './redux'
